backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-migrate
python-dotenv
psycopg2-binary
pyjwt
flask-bcrypt
google-auth
gevent
requests
PyPDF2
langchain==0.1.4
openai
tiktoken
langchain_openai
sentence-transformers
faiss-cpu